Rename `retrieve` to `retrieve_or_build`
This is to make it clearer that its functionality is not only to
retrieve a link based on the given params, but also to instantiate a new
link with the given params if one is not present in the database. Tests
have also been refactored and added accordingly to make this explicit.