Fix link spec local authority bug
Fixing a linting error in the `existing_homepage` method on the Link model
uncovered a bug in the link model spec. When running the test, the link factory
would create an associated Local Authority. While the Local Authority created
by the link factory had the same homepage URL as the Local Authority that was
created before it, its status was set to nil. This wasn't an issue when the
`existing_homepage` method looked up the first local authority with the given
homepage URL; now it has changed the wrong local authority is being returned
and therefore the wrong status is being assgined to the link. We wouldn't
expect multiple local authorities to have the same homepage URL so we shouldn't
have a duplicate local authority being created by the link factory; as such we
assign it the local authority that was originally created, which resolves the
issue.