Rename 'find_by' methods to 'lookup_by'
Rubocop has an issue (see [1]) with `find_by_attribute` methods. This app does
have `find_by` methods which are defined on the models as the queries are more
complex than a simple attribute lookup, but they are failing the linter because
of how they are named. Instead of littering the app with disable rubocop
statements, it feels simple to rename the methods to `lookup_by`.

[1] https://rails.rubystyle.guide/#find_by