Use 'update!' to surface save errors
Previously we switched from 'update_attribute' to 'update' to fix
linting issues, but this carries a risk of silently failing to save
changes, since 'update' will run validations. This switches all uses
of 'update' to 'update!' to ensure we become aware of any issues in
case they are not covered by the tests.