Document setup instructions and move docs into docs/
We now use govuk-docker to run the application locally.

I've moved some documentation into the docs/ folder, for
consistency, as there was an odd split of what was in the folder
and what was in the README.

I've removed the link to the Production environment, as other
app READMEs don't have this and it doesn't fit the heading of
'live examples'.