Make import script robust to invalid URLs
When a CSV contains a 'New URL' that is invalid, we currently
show the default Rails 'Something went wrong' page, which is a
poor user experience as it doesn't give any clues as to what went
wrong. It also didn't log the error in Sentry.

We now:

- Log the error in Sentry, along with detailed debug information
- Capture the error and prevent the Rails default error page
- Present a custom error to the user, giving them the exact line
  the issue happens on. If there are multiple issues, we show
  all of the errors, separated by line breaks (<br>).

Trello: https://trello.com/c/5IFlmQQQ/2039-3-improve-local-links-manager-error-handling-for-the-csv-upload-feature