Add local links for Test and Trace scheme
What

We need to add the local authority links for the test and trace scheme
self isolation payments.

The rake task takes 3 params: LGSL code, LGIL code and a filename. The
file to which the filename points to must be a CSV file with 2 columns.
It must have a header row of "slug" and "url". It is expected that this
file be available locally, in the /tmp directory for example.

Why

The test and trace scheme is newly created.