Remove need for REDIS_HOST and REDIS_PORT env vars
This is one of the few GOV.UK projects that uses this old approach to
configuring Redis - whereas most of the other ones use the approach
supported natively by the Redis gem [1] of using REDIS_URL. Removing
this will allow us to remove the need for REDIS_HOST and REDIS_PORT env
vars in our architecture and also allows this code to be simplified.

It's worth noting that Redis Namespace was not actually having any
effect on this project so I have removed the gem. The `namespace`
argument as part of a Redis config is a Sidekiq feature, not a Redis
one. There is already a namespace set manually as part of the code in
DistributedLock.

[1]: https://github.com/redis/redis-rb