Add rake task to enable or create services
This adds a rake task that can be used to enable an existing service or
create a dummy service (if not defined by LGA). Enabling a service
involves making them live and creating the service interaction. This
used the "providing information" interaction by default (i.e. argument
unspecified). This superseded the migrations used to enable services and
makes it easier to enable new services. Migration should probably not be
used for defining data added to the database.

Co-authored-by: Graham Lewis <gclssvglx@gmail.com>