Add functionality to duplicate a service
When the find-covid-19-lateral-flow-test-site service was originally created, it used a dummy LGSL code (100001). We now have a real LGSL code for this service (1827) and would like to use this LGSL code instead of the dummy code.

As it is not possible to change the LGSL code of a service via the user interface, and due to the need to keep the service in-sync between LocalLinksManager and Publisher so that the service is not interrupted - it is safer to create a duplicate of the original service, then switch Publisher over to the duplicate service before finally removing the original service.

This commit adds the ability to duplicate an existing service.