Add functionality to update a service label and slug
Both the label and the slug of a service have a uniqueness constraint on them within the model.  So creating or updating a service with the same label and/or slug as an existing service, will fail.

Changing either the label or slug is not possible via the user interface.

This commit adds the ability to change the label and/or the slug of a service.  Aside from giving the ability to change either or both of these attributes, it can also be used on those occasions when a duplicate of a service may be required (for example, when changing the LGSL code of a service).