Add functionality to destroy a service
Retired or no longer used services are not currently automatically removed from the system when no longer required. It is also not currently possible to remove them via the user interface. So, without the ability to clean them up - the system just accumulates services over time.

This commit adds the ability to remove a service from the system.

In Rails, all models that inherit from ActiveRecord have a method named 'destroy!' available that is responsible for removing a record and (potentially) any/all associated child records. It would be ideal to use the 'destroy!' method for removing a service.  However, a service has an association with one or more ServiceTier records. These records indicate what level of local/regional or national authority the service is available in and are not automatically destroyed when the parent Service is removed. This causes a call to 'destroy!' on a service to fail. It is felt that these ServiceTier records should be removed with the parent, as ServiceTier records serve no purpose without a parent Service. To do this the 'has_many :service_tiers, dependent: :restrict_with_error' has been changed to 'has_many :service_tiers, dependent: :destroy'.