Consolidate tasks to import service links from CSV
- The two deleted tasks were very similar. The difference was that links.rake
found a LocalAuthority based on its 'name' property and covid19_test_and_trace_scheme_links.rake
found a LA based on its 'slug' property.
- Finding by 'slug' feels more reliable as a slug is less likely to have spaces or characters
that might cause problems in the CSV data.

The encoding has been set to `encoding: 'bom|utf-8'`.
The reason for this is explained [here](https://stackoverflow.com/a/50336647) in which
someone "discovered that saving from an .xlsx workbook in MS Excel, to the .csv is what caused those strange characters to appear". I did the same thing and ended up getting confused when it wasn't working due to bytes being added which are unseen to the human eye... this was fixed when re-writing the data in a text editor, but it might save someone else some pain and also makes it more generic.

These characters being added is a result of [Byte Order Mark (BOM) for UTF-8 encoding](https://estl.tech/of-ruby-and-hidden-csv-characters-ef482c679b35).

To run the task (stolen from Graham's original commit):

The rake task takes 3 params: LGSL code, LGIL code and a filename. The
file to which the filename points to must be a CSV file with 2 columns.
It must have a header row of "slug" and "url". It is expected that this
file be available locally, in the /tmp directory for example.