Update README.md
Adds a hyperlink which was lost by mistake in the previous commit.

As before, this update is primarily a learning process for me.  The reason for creating a new commit each time rather than amending my previous commits is that I don't know how to do that, and don't plan to learn at this early stage in my Git journey.