Add a user interface to edit authority homepage URLs
Currently local authority homepage URLs are updated by running a rake
task. This limits these changes to developers only and results in an
unnecessary time consuming process to make such changes.

This adds a basic user interface so the changes can be made by anyone
with access to the application.

After this has been deployed, the rake task and related documentation
can be removed.