Add ability to find local authority by local custodian code
This is required for frontend apps to be able to switch from Mapit
to Locations API; the latter returns the 'local custodian code'
associated with the local authority a given postcode resides in.
We need to be able to find a local authority by its local
custodian code.

Trello: https://trello.com/c/S2gw5MAZ/2933-5-add-methods-for-local-links-manager-to-lookup-local-authority-by-local-custodian-code