Expose 'slug' in API response for local authority presenter
There was little point in exposing this before, as the only way
of searching for a local authority was to pass its slug to the
API. Now, however, it is possible to look up a local authority by
its local custodian code, so the upstream app may not actually
know the slug.

Various parts of Local Links Manager require the local authority
slug as a parameter ([1], [2], [3]). Rather than change all of
these to take a local custodian code as an optional parameter, it
is more pragmatic to simply expose the slug in the local authority
presenter. Upstream apps can then choose to use the returned slug
in subsequent requests to Local Links Manager's API.

[1]: https://github.com/alphagov/local-links-manager/blob/35414db8a7d51c3c3c54c7413ab6e9642b709dbc/app/lib/local_links_manager/check_links/link_status_requester.rb#L19
[2]: https://github.com/alphagov/local-links-manager/blob/35414db8a7d51c3c3c54c7413ab6e9642b709dbc/app/controllers/local_authorities_controller.rb#L22
[3]: https://github.com/alphagov/local-links-manager/blob/d846427f11ac54780cca4c56ce9aa36740eb077e/app/controllers/links_controller.rb#L58