Give 400 response if local_custodian_code & authority_slug present
We don't want a situation where the upstream app can pass a slug
for one local authority and a local custodian code for a different
local authority - the request is ambiguous and should not be
actioned.

Whilst we could perform the lookup and only return a 400 if both
parameters resolve to different local authorities, that is
significantly more complicated than simply returning immediately
if both parameters are provided. In any case, we do not want to
encourage the passing of an unnecessary 2nd parameter in any API
requests.