Add Local Authorities CSV file
We need to remove our dependency on Mapit for keeping local
authority data up to date. This means removing the
`import:local_authorities:import_all` rake task, which pulls in
local authority data from Mapit.

Whilst we could replace this with Rake tasks to add or update
individual local authorities, this introduces risk, as developers
will be updating records by hand, and we'll no longer have the
single source of truth that Mapit provides. In the event that we
lost our database and backups, it would be difficult to get all the
configuration back.

Given that there is nothing sensitive in this information, and the
quantity of local authorities is manageable, it makes sense to
store the data in version control. We then have a data source we
can import in an idempotent way.

We'll use this CSV in a rake task in a later commit. The CSV was
generated by running the following in a Rails console:

```
rows = LocalAuthority.all.sort.pluck(:id, :gss, :snac, :local_custodian_code, :tier_id, :parent_local_authority_id, :slug, :country_name, :homepage_url, :name).map do |cols|
  cols[cols.count-1] = "\"#{cols.last}\""
  cols.join(",")
end
rows.each { |row| puts row } && nil
```

Trello: https://trello.com/c/ta3NoTNg/2947-replace-mapit-powered-rake-tasks-in-local-links-manager-5