Test that data CSV is valid
This checks that the CSV has the right number of values in each
row, and that each value is in the format we expect. It also
checks that the headings are in the order we expect.

The regexes used in these tests are for sense-checking only. For
example, checking that the URL field begins with `http` is
a sufficient sense-check, rather than finding a water-tight URL
validator, which should be the responsibility of the model.

In writing these tests, I noted a few quirks with the current
production data, which I will correct in the next few commits.