Copy mapit.json fixture data into local-authorities.csv fixture
We're about to switch over from Mapit import to CSV import. We
want to use the same tests, so it makes sense to base it off the
same kind of fixture data, albeit a CSV rather than the JSON file:
https://github.com/alphagov/local-links-manager/blob/2293da6fe6baf88bcf6666fb844d227bbc02e698/spec/lib/local-links-manager/import/fixtures/mapit.json

All the data is the same except for the local custodian code,
which I've made up arbitrarily, and the ID, which I've reused the
Mapit ID for. It shouldn't make a difference in practice.