Tolerate lack of /usr/local/bin/notify_passive_check.
The /usr/local/bin/notify_passive_check binary doesn't exist when
running on Kubernetes, so make it optional.

Intentionally not making this configurable, because to do so would be
overkill at this point. If the job were that important it'd have
freshness monitoring anyway, and that would still fire if the push to
Icinga were to stop working for any reason.