Add succeeded_by field to local authorities
This change is needed to handle the situation where
future merges are recorded but they're into a council
that isn't the current parent (for instance Cumbria,
which is being split into two Unitary Authorities in
April 2023 and the District Councils shared between
them. They'll now show Cumbria up to the switch point,
and the relevant Unitary Authorities after that.